feat: separate out gallery and details into standalone sc
